+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Release Notes
 
 - add 'on the block' indicator
+- fix players being moved or removed during nomination
+- add vue linter
+- use "Exile" rather than "Banishment" for exiles
+- added global animation toggle for better performance
+- added record vote history toggle to session menu, and clear vote history button
+- add support for custom Fabled characters
+
+---
 
 ### Version 2.12.0
 - tweak reference sheet to better fit screen in single column layout
 - add warning icon overlay for setup roles on character assignment modal
-- added Heretic and Marionette plus King and Choirboy to list of available characters
+- added Heretic and Marionette plus King/Choirboy and the Gangster to list of available characters
 
 ---